# API-RESTful/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Podcast-Finder API",
        "description": "Documentation de l'API Podcast-Finder",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000/api",
            "description": "Local API"
        }
    ],
    "paths": {
        "/api/podcasts/{podcast_id}/episodes": {
            "get": {
                "tags": [
                    "Episodes"
                ],
                "summary": "Liste des épisodes d'un podcast",
                "operationId": "c69fe57d6278ec456b66f447e361abc5",
                "parameters": [
                    {
                        "name": "podcast_id",
                        "in": "path",
                        "description": "ID du podcast",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Liste des épisodes",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Episode"
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Podcast introuvable"
                    },
                    "401": {
                        "description": "Non authentifié"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/podcasts": {
            "get": {
                "tags": [
                    "Podcasts"
                ],
                "summary": "Liste de tous les podcasts",
                "operationId": "28abe6e0a78fe6785659a51dde99679e",
                "responses": {
                    "200": {
                        "description": "Liste des podcasts avec leurs animateurs",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Podcast"
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Non authentifié"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "schemas": {
            "Episode": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "title": {
                        "type": "string",
                        "example": "Épisode 1"
                    },
                    "description": {
                        "type": "string",
                        "example": "Description"
                    },
                    "audio": {
                        "type": "string",
                        "example": "https://res.cloudinary.com/.../audio.mp3"
                    },
                    "podcast_id": {
                        "type": "integer",
                        "example": 1
                    }
                },
                "type": "object"
            },
            "Podcast": {
                "title": "Podcast",
                "description": "Modèle de podcast",
                "required": [
                    "titre",
                    "categorie",
                    "description"
                ],
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "titre": {
                        "type": "string",
                        "example": "Mon Podcast Tech"
                    },
                    "categorie": {
                        "type": "string",
                        "example": "Technologie"
                    },
                    "description": {
                        "type": "string",
                        "example": "Un podcast sur les nouvelles technologies"
                    },
                    "image": {
                        "type": "string",
                        "example": "https://cloudinary.com/image.jpg"
                    },
                    "user_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2024-01-15T10:30:00Z"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2024-01-15T10:30:00Z"
                    },
                    "user": {
                        "$ref": "#/components/schemas/User"
                    },
                    "episodes": {
                        "description": "Liste des épisodes du podcast",
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Episode"
                        }
                    }
                },
                "type": "object"
            },
            "User": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "first_name": {
                        "type": "string",
                        "example": "Housna"
                    },
                    "last_name": {
                        "type": "string",
                        "example": "Fathi"
                    },
                    "email": {
                        "type": "string",
                        "example": "housna@example.com"
                    },
                    "role": {
                        "type": "string",
                        "example": "animateur"
                    }
                },
                "type": "object"
            }
        }
    },
    "tags": [
        {
            "name": "Episodes",
            "description": "Episodes"
        },
        {
            "name": "Podcasts",
            "description": "Podcasts"
        }
    ]
}